vm: move DVM template specific code into separate mixin
No functional change.
