+++ rpm_spec/core-dom0.spec.in
@@ -249,6 +249,7 @@ fi
 %dir %{python3_sitelib}/qubes/vm/mix/__pycache__
 %{python3_sitelib}/qubes/vm/mix/__pycache__/*
 %{python3_sitelib}/qubes/vm/mix/__init__.py
+%{python3_sitelib}/qubes/vm/mix/dvmtemplate.py
 %{python3_sitelib}/qubes/vm/mix/net.py
 
 %dir %{python3_sitelib}/qubes/storage
